content_pipeline/requirements.txt
openai
requests
python-slugify
gspread
google-auth